6 --- data truncated for column equipPrice

+++ src/main/java/com/example/prog1/pattern/decorator/Price.java
@@ -1,7 +1,7 @@
 package com.example.prog1.pattern.decorator;
 
 public interface Price {
-    Double getPrice();
+    Integer getPrice();
     /**
      * partiamo da un totale che è 0 perchè non abbiamo nulla nel carrello, aggiungendo cose "decoriamo" il prezzo con
      * nuovi dati. */
